requirements.txt
fastapi>=0.100.0
uvicorn>=0.25.0
databricks-sdk>=0.40.0
pydantic>=2.0
pydantic-settings>=2.0
apscheduler>=3.10.4
emails>=0.6
Jinja2>=3.1.2
mcp>=1.0.0
httpx>=0.24.0
